Lager script for å koble proxy mot nginx i docker-compose for frontendutvikling
