+++ package.json
@@ -115,7 +115,7 @@
     "patch:init": "node ./scripts/patchinit.js",
     "patch:env": "node ./scripts/patchenv.js",
     "patch:watch": "node ./scripts/patchwatch.js",
-    "patch:javalocal": "node ./scripts/patchjavalocal.js",
+    "patch:nginxlocal": "node ./scripts/patchnginxlocal.js",
     "analyze": "source-map-explorer build/static/js/main.*",
     "start:js": "react-scripts start",
     "build:js": "react-scripts build",
@@ -125,7 +125,7 @@
     "cypress:open": "cypress open",
     "cypress:run": "cypress run",
     "start": "npm-run-all -s patch:init patch:env echo:env -p start:css start:js",
-    "start:javalocal": "npm-run-all -s patch:init patch:env patch:javalocal echo:env -p start:css start:js",
+    "start:nginxlocal": "npm-run-all -s patch:init patch:env patch:nginxlocal echo:env -p start:css start:js",
     "start:kubefwd": "npm-run-all -s patch:init patch:env patch:kubefwd echo:env -p start:css start:js",
     "build": "npm-run-all -s patch:init patch:env echo:env -p build:css build:js",
     "watch": "npm-run-all -s patch:init patch:env patch:watch echo:env -p start:css build:watch",
