+++ src/setupProxy.js
@@ -1,8 +1,17 @@
 const { createProxyMiddleware } = require("http-proxy-middleware");
 
 module.exports = function setupProxy(app) {
-  app.use(createProxyMiddleware("/api", { target: "http://localhost:3002/" }));
-  app.use(createProxyMiddleware("/graphql", { target: "http://localhost:3002/" }));
-  app.use(createProxyMiddleware("/frontendlogger", { target: "http://localhost:3002/" }));
-  app.use(createProxyMiddleware("/flyt", { target: "http://localhost:8088/" }));
+  app.use(
+    createProxyMiddleware(["/api", "/graphql", "/frontendlogger"], {
+      target: `http://localhost:${process.env.REACT_APP_LOCAL_API_PORT}/`,
+    })
+  );
+  app.use(
+    createProxyMiddleware("/trygdeavtale-flyt", {
+      target: "http://localhost:8088/",
+      pathRewrite: {
+        "^/trygdeavtale-flyt": "/flyt",
+      },
+    })
+  );
 };
